+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## unreleased
+
+- Allow custom encryptor per attribute.
+
 ## v0.1.0
 
 - Initial release.